feat: add update non-volunteer profile for host and org manager method, add view host and org manager profile method

+++ src/main/java/com/sep490/g28/hvh/be/service/impl/VolunteerServiceImpl.java
@@ -1,8 +1,7 @@
 package com.sep490.g28.hvh.be.service.impl;
 
 import com.sep490.g28.hvh.be.auth.CurrentUserProvider;
-import com.sep490.g28.hvh.be.constant.ERole;
-import com.sep490.g28.hvh.be.constant.EVolunteerVerificationStatus;
+import com.sep490.g28.hvh.be.constant.*;
 import com.sep490.g28.hvh.be.dto.volunteer.request.RegisterVolunteerAccountRequest;
 import com.sep490.g28.hvh.be.dto.volunteer.request.UpdateVolunteerProfileRequest;
 import com.sep490.g28.hvh.be.dto.volunteer.request.VolunteerRegistrationVerifyRequest;
@@ -498,19 +497,40 @@ public class VolunteerServiceImpl implements VolunteerService {
     }
 
     @Override
-    public void updateVolunteerProfile(UpdateVolunteerProfileRequest request) {
+    public UpdateVolunteerProfileResponse updateVolunteerProfile(UpdateVolunteerProfileRequest request) {
         UUID volunteerId = currentUserProvider.getId();
 
         Volunteer volunteer = volunteerRepository.findById(volunteerId)
                 .orElseThrow(() -> new AppException(VolunteerErrorCode.VOLUNTEER_NOT_EXISTED));
 
-        if (volunteer.getAvatarUrl() != null && !volunteer.getAvatarUrl().isEmpty()) {
-            //delete exist avatar image
-            CompletableFuture<Void> avatarFuture =
-                    storageService.deleteFileAsync(volunteer.getAvatarUrl());
+        String newAvatarUploadUrl = null;
+        if(request.getAvatarExtension() != null) {
+            if (volunteer.getAvatarUrl() != null && !volunteer.getAvatarUrl().isEmpty()) {
+                //delete exist avatar image
+                CompletableFuture<Void> avatarFuture =
+                        storageService.deleteFileAsync(volunteer.getAvatarUrl());
+
+                try {
+                    CompletableFuture.allOf(avatarFuture).join();
+                } catch (CompletionException e) {
+                    Throwable cause = e.getCause();
+                    if (cause instanceof AppException ae && ae.getHttpStatus().value() == 400) {
+                        //todo: this case is the file not exist in sb (only for test) change later, need to have picture to approve
+                    } else {
+                        throw (RuntimeException) e.getCause(); // propagate, transaction fail
+                    }
+                }
+            }
+
+            //get signed URL of file
+            String newAvatarPath = storagePathGenerator.volunteerAvatar(volunteerId, request.getAvatarExtension());
+
+            CompletableFuture<String> newAvatarFuture =
+                    storageService.getUploadUrlAsync(newAvatarPath);
 
             try {
-                CompletableFuture.allOf(avatarFuture).join();
+                CompletableFuture.allOf(newAvatarFuture).join();
+                newAvatarUploadUrl = newAvatarFuture.join();
             } catch (CompletionException e) {
                 Throwable cause = e.getCause();
                 if (cause instanceof AppException ae && ae.getHttpStatus().value() == 400) {
@@ -520,7 +540,36 @@ public class VolunteerServiceImpl implements VolunteerService {
                 }
             }
 
+            volunteer.setAvatarUrl(newAvatarPath);
         }
+
+        EEmployStatus employStatus =
+                (request.getEmployStatus() == null || request.getEmployStatus().isBlank())
+                        ? null
+                        : EEmployStatus.valueOf(request.getEmployStatus());
+
+        EEducationLevel educationLevel =
+                (request.getEducationLevel() == null || request.getEducationLevel().isBlank())
+                        ? null
+                        : EEducationLevel.valueOf(request.getEducationLevel());
+
+        volunteer.setFullName(request.getFullName());
+        volunteer.setBio(request.getBio());
+        volunteer.setGender(request.isGender());
+        volunteer.setDob(request.getDob());
+        volunteer.setAddress(request.getAddress());
+        volunteer.setDetailAddress(request.getDetailAddress());
+        volunteer.setEmployStatus(employStatus);
+        volunteer.setWorkAddress(request.getWorkAddress());
+        volunteer.setEducationLevel(educationLevel);
+        volunteer.setSid(request.getSid());
+        volunteer.setDeviceId(request.getDeviceId());
+
+        volunteerRepository.save(volunteer);
+
+        return UpdateVolunteerProfileResponse.builder()
+                .avatarUploadUrl(newAvatarUploadUrl)
+                .build();
     }
 
     //convert name to valid username
